Updated Nordea to use the non-mobile page instead. The problem with the mobile page is that it shows "Kortköp" for everything, not where the purchase originated. The non-mobile page does not have this issue.
Note this update does not include support for loans, credit cards or funds. Only regular accounts. This as I do not have any of the other three personally and don't know how the pages look.

Also of note, Nordea is a pain with its variables that need to be sent between pages. Send one too few or too many and they send bad results.

Signed-off-by: Johan Sköld <johan@skold.cc>