Don't use Server Name Indication when refreshing certificates
It looks like some servers return different certificates depending on
whether SNI is set or not. Bankdroid doesn't seem to set SNI when
connecting to the banks, so the script refresh_bank_certificates
shouldn't use SNI either or there will be certificate verification
failures.

This fixes certificate verification for Östgötatrafiken.

I've refreshed all bank certificates with this change and only a few
certificates changed. I've verified that Östgötatrafiken works. The
other affected banks look disabled (except for Meniga bank, but do we
know if that bank is working?)