Enable all non-violated PMD rules
And add a script for updating the list.

Note that I haven't actually fixed anything in here, and there are some
kind-of-scary-looking suppressions (like "EmptyIfStmt").

Fixing those would be another exercise though, the only point of this
change is really to document what we're violating and make sure we don't
add to that list.