Add bank-disabled warning to TransactionsActivity
If a bank is disabled, and you click on the widget, with this change in
place you now get a warning about the bank being disabled and what you
can do about it.

This is part of fixing #610.

To be able to work on this I first fixed all warnings in the two files
I had to change. To remove the suppressions for those warnings so that
they don't come back, I fixed all other instances of those problems as
well.