Show warning for disabled banks
Before this change, in the banks list in the main UI, disabled banks
were marked with a warning icon.

This change adds to that warning icon a text describing what's actually
going on and what the user can do about it.

Before starting to work on this I fixed all Android Lint reported
warnings in the .java and .xml files I had to change, and re-enabled
those warnings.

One of those warnings was about using sp rather than dp for text. I
suppressed that warning for the widgets, but heeded it in all other
places. This results in a UI that better adapts to the user's font size
preferences.

The UI changes I have tested with both "small" and "huge" font size,
and it looks fine.